Refactor MakeConvexHull() to not depend on Shape
The core functionality has been broken into its own translation unit, one
that does not depend on Shape. The previous API is maintained for
convenience.

This is a precursor to allowing the Mesh and Convex shape to compute their
own convex hulls upon demand.
